### 2025-05-17 #8228 @editorialbot
[PRE REVIEW]: PileupCaller: A command-line tool to sample genotypes from low-coverage sequencing data of ancient DNA
**Submitting author:** @stschiff (<a href="http://orcid.org/0000-0002-1017-9150">Stephan Schiffels</a>)
**Repository:** https://github.com/stschiff/sequenceTools
**Branch with paper.md** (empty if default branch): writing_joss_paper
**Version:** v1.6.0.0
