### 2025-06-19 #8228 @stschiff
Thanks. Will compile some suggestions later.
