### 2025-07-01 #8228 @mahfuz05062
> Hi [@dkoptekin](https://github.com/dkoptekin) [@uqrmaie1](https://github.com/uqrmaie1) [@TCLamnidis](https://github.com/TCLamnidis) - are any of you interested and have the bandwidth to review this article? Thanks!

Hi @dkoptekin @uqrmaie1 @TCLamnidis - checking once more to see if you have any bandwidth to take on the review. Thanks!
