### 2025-07-22 #8634 @editorialbot
[REVIEW]: PileupCaller: A command-line tool to sample genotypes from low-coverage sequencing data of ancient DNA
**Submitting author:** @stschiff (<a href="http://orcid.org/0000-0002-1017-9150">Stephan Schiffels</a>)
**Repository:** https://github.com/stschiff/sequenceTools
**Branch with paper.md** (empty if default branch): master
**Version:** v1.6.0.0
