### 2025-10-29 #8634 @mahfuz05062
Hi @josecediel25 checking one more time if you have the chance to finish your checklist. If I don't hear from your side in about a week, I will assume things are mostly fine and will initiate the next steps from my side.
