### 2026-01-16 286bcca
added ubuntu arm runner, added all three tools to release workflow
