### 2026-01-16 ddce972
fixed problem with set path step
