+++ .github/workflows/release.yaml
@@ -52,29 +52,34 @@ jobs:
           test: false
           stack-build-arguments: --copy-bins --ghc-options="-O2"
 
-      - name: Set binary path name
-        id: binarypath
-        run: for TOOL in pileupCaller vcf2eigenstrat genoStats; do
+      - name: Set binary path names
+        id: binarypaths
+        run: |
+            binaries=()
+            for TOOL in pileupCaller vcf2eigenstrat genoStats; do
               if [ "$RUNNER_OS" == "Windows" ]; then
                   newEXE="$TOOL-$RUNNER_OS.exe"
               else
                   newEXE="$TOOL-$RUNNER_OS-$RUNNER_ARCH"
               fi
               currentEXE="${{ steps.stack.outputs.local-bin }}/$TOOL"
-              mv $currentEXE $newEXE
-              echo "BINARY_PATH=$newEXE" >> $GITHUB_OUTPUT
+              mv "$currentEXE" "$newEXE"
+              binaries+=("$newEXE")
             done
+            # join with commas for release-action
+            binary_paths=$(printf "%s," "${binaries[@]}")
+            binary_paths=${binary_paths%,}
+            echo "BINARY_PATHS=$binary_paths" >> "$GITHUB_OUTPUT"
         shell: bash
 
-      - name: Upload Release Asset
-        id: upload-release-asset
+      - name: Upload Release Assets
         uses: ncipollo/release-action@v1
         with:
           name: Release ${{ github.ref_name }}
           draft: true
           allowUpdates: true
           artifactErrorsFailBuild: true
-          artifacts: ${{ steps.binarypath.outputs.BINARY_PATH }}
+          artifacts: ${{ steps.binarypaths.outputs.BINARY_PATHS }}
           artifactContentType: application/octet-stream
 
   # build_centos_artifact:
