### 2026-01-16 ca3047f
reactivated docker build for testing
