### 2026-01-16 2b69ea2
reactivating docker build
