### 2026-01-28 #8634 @stschiff
I've completed the post-review checklist from above.

The new release of the software is [version v1.6.0.0](https://github.com/stschiff/sequenceTools/releases/tag/v1.6.0.0)
The release also has a Zenodo DOI: [10.5281/zenodo.18400553](https://doi.org/10.5281/zenodo.18400553)
