+++ .github/workflows/release.yaml
@@ -118,9 +118,20 @@ jobs:
       - name: Checkout repo
         uses: actions/checkout@v4
 
-      - name: Build Docker image
-        run: docker build -t linux -f .github/workflows/Dockerfile.centos_aarch64 .
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v3
 
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Build Docker image
+        run: |
+          docker buildx build \
+            --platform linux/arm64 \
+            -t linux \
+            -f .github/workflows/Dockerfile.centos_aarch64 \
+            --load .
+     
       - name: Create container
         run: docker create --name linuxcontainer_aarch64 linux
 
