### 2026-02-16 #8634 @mahfuz05062
Hi @stschiff , I went through the article and it looks good. I have one question though: I see that in the archive (on Zenodo) there are 3 different contributors listed, however, only one author on the article. Is this expected?

Also, please update the citation for Genome-wide patterns of selection in 230 ancient eurasians (https://www.nature.com/articles/nature16152). The listed DOI in this article is "https://doi.org/10.1038/nature16152" whereas the current citation uses a "doifinder" link, which is unusual.
