+++ .github/workflows/release.yaml
@@ -110,41 +110,3 @@ jobs:
           artifacts: pileupCaller-conda-linux
           artifactContentType: application/octet-stream
 
-  build_centos_artifact_aarch64:
-    needs: [create_release]
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v4
-
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v3
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Build Docker image
-        run: |
-          docker buildx build \
-            --platform linux/arm64 \
-            -t linux \
-            -f .github/workflows/Dockerfile.centos_aarch64 \
-            --load .
-     
-      - name: Create container
-        run: docker create --name linuxcontainer_aarch64 linux
-
-      - name: Copy executable
-        run: docker cp linuxcontainer_aarch64:/root/.local/bin/pileupCaller pileupCaller-conda-linux-aarch64
-
-      - name: Upload Release Asset
-        id: upload-release-asset
-        uses: ncipollo/release-action@v1
-        with:
-          name: Release ${{ github.ref_name }}
-          draft: true
-          allowUpdates: true
-          artifactErrorsFailBuild: true
-          artifacts: pileupCaller-conda-linux-aarch64
-          artifactContentType: application/octet-stream
