### 2026-02-18 cbce2b6
updated conda docker to build all three executables
