### 2026-02-18 #8634 @stschiff
Ah thanks for catching that. No that was a mistake on Zenodo. I think I might have received very minor Pull Requests from the two additional authors at some point in recent years, but I would consider their contribution to be negligible. I can't remember what that was, maybe just a note in the README or so. Nothing code-related. I just removed the two from the Zenodo-archived version.
