+++ paper/paper.bib
@@ -7,8 +7,6 @@
   url = {https://github.com/stschiff/sequence-formats}
 }
 
-
-
 @misc{eigensoft_github,
   author = "Mah, Matthew and Patterson, Nick and Price, Alkes",
   title = {Eigensoft},
@@ -50,7 +48,6 @@
   language  = "en"
 }
 
-
 @ARTICLE{Price2006,
   title     = "Principal components analysis corrects for stratification in
                genome-wide association studies",
@@ -63,8 +60,9 @@
   pages     = "904--909",
   month     =  "23~" # jul,
   year      =  2006,
-  url       = "http://www.nature.com/doifinder/10.1038/ng1847",
+  url       = "https://www.nature.com/articles/ng1847",
   file      = "All Papers/P/Price et al. 2006 - Principal components analysis corrects for stratification in genome-wide association studies.pdf",
+  doi       = "10.1038/ng1847",
   issn      = "1061-4036"
 }
 
@@ -83,8 +81,8 @@
                for Cancer Research} and {Pennsylvania State University} and
                {Rutgers Cancer Institute of New Jersey} and {Tel-Aviv
                University} and {Jackson Laboratory for Genomic Medicine} and
-               {Thermo Fisher Scientific} and University of California, San
-               Francisco and {University of Chicago} and {University College
+               {Thermo Fisher Scientific} and {University of California, San
+               Francisco} and {University of Chicago} and {University College
                London} and {University of Geneva} and {University of
                Montr\'{e}al} and {University of North Carolina at Chapel Hill}
                and {University of Puerto Rico} and {University of Texas Health
@@ -93,7 +91,7 @@
                {Cold Spring Harbor Laboratory} and {European Molecular Biology
                Laboratory} and {Leiden University Medical Center} and {Louisiana
                State University} and {McDonnell Genome Institute at Washington
-               University} and University of California, San Diego and
+               University} and {University of California, San Diego} and
                {University of Maryland School of Medicine} and {University of
                North Carolina at Charlotte} and {University of Oxford} and
                {University of Texas MD Anderson Cancer Center} and {University
@@ -108,20 +106,20 @@
                University} and {European Molecular Biology Laboratory, European
                Bioinformatics Institute} and {University of Queensland} and
                {Virginia Bioinformatics Institute} and {Data coordination center
-               group} and {Illumina} and University of California, Santa Cruz
+               group} and {Illumina} and {University of California, Santa Cruz}
                and {Samples and ELSI group} and {Sample collection} and
                Colombians in Medell\'{\i}n, Colombia Clm and {Finnish in Finland
                (FIN)} and {Iberian Populations in Spain (IBS)} and {Puerto
                Ricans in Puerto Rico (PUR)} and {African Caribbean in Barbados
                (ACB)} and {Esan in Nigeria (ESN)} and {Gambian in Western
-               Division -- Mandinka (GWD)} and Kinh in Ho Chi Minh City, Vietnam
-               (khv) and Punjabi in Lahore, Pakistan (pjl) and {Scientific
-               management} and {Writing group} and {National Eye Institute, NIH}
-               and {British from England and Scotland (GBR)} and {Han Chinese
-               South (CHS)} and {Bengali in Bangladesh (BEB)} and Chinese Dai in
-               Xishuangbanna, China (cdx) and {Indian Telugu in the UK (ITU) and
-               Sri Lankan Tamil in the UK (STU)} and {Mende in Sierra Leone
[…]
