### 2026-02-18 #8634 @stschiff
OK, I've updated the citations. I have pushed everything now to the master branch. Is that OK or should I update the dedicated branch as well?
