### 2026-02-18 #8634 @stschiff
Can we just edit the submission to indicate that the default branch is the one with the most recent version of the paper?
