### 2026-03-15 #8634 @mahfuz05062
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.1093/bioinformatics/btr330 is OK
> - 10.1093/gigascience/giab008 is OK
> - 10.1038/s43586-020-00011-0 is OK
> - 10.7554/elife.98317 is OK
> - 10.1038/s41597-024-03031-7 is OK
> - 10.1101/gr.276728.122 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Sequence-Formats
> - No DOI given, and none found for title: Eigensoft
> - No DOI given, and none found for title: Principal components analysis corrects for stratif...
> - No DOI given, and none found for title: A global reference for human genetic variation
> - No DOI given, and none found for title: Genomics in the cloud: using Docker, GATK, and WDL...
> - No DOI given, and none found for title: Massive migration from the steppe was a source for...
>
> ❌ MISSING DOIs
>
> - 10.1038/nature16152 may be a valid DOI for title: Genome-wide patterns of selection in 230 ancient E...
>
> ❌ INVALID DOIs
>
> - None
> ```

The link for the "Missing DOIs" document points to the right document.
