### 2026-03-15 #8634 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
