### 2026-03-15 #8634 @Kevin-Mattheus-Moerman
@stschiff As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which require your attention:

@stschiff GitHub does not automatically detect your license. Likely because the text was slightly edited from the standard text. Can you confirm your license type for me please? Is it the [BSD-3-clause](https://opensource.org/license/BSD-3-clause) type? If so perhaps you can edit your license text to match the standard text so that automated systems like ours recognise your license.
->  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

- [ ] Review completed?
It appears @josecediel25 did not complete the review as several boxes are not ticked. I have raised the issue of the license and the community guidelines, so they should be addressed shortly. However, @mahfuz05062 can you comment on the unticked boxes relating to functionality and documentation? Can you verify these before we proceed?

- [ ] Copyedit pass/checked for typos

### Paper text
- [ ] `seuqencing` - > `sequencing
- [ ] `accellerating` -> `accelerating`
- [ ] Your paper currently lacks an acknowledgements section (e.g. to acknowledge funding). This may be intentional, and the section is not required, however if you do feel you need one please add one now.
