### 2026-04-01 c6df79a
added contributing guide and Code of Conduct
