### 2026-04-01 d54d02d
fixed typos and added acknowledgments
