+++ paper/paper.md
@@ -17,11 +17,11 @@ bibliography: paper.bib
 
 # Summary
 
-Next generation sequencing data is ubiquitous in medical and biological sciences. It has also become the primary tool in archaeogenetics, where ancient DNA is extracted from archaeological organic (often human skeletal) material, processed into DNA sequencing libraries and then sequenced [@Orlando2021]. As a testimony to the rapid and accellerating growth of the field, we today have close to ten thousand published ancient human genomes available in the public record [@Schmid2024; @Mallick2024], and many smaller datasets of other organisms. A key step in processing raw sequencing data is the estimation of genotypes at specific variable positions along the genome. Such positions are often pre-selected because they are informative about ancestry or of particular biological relevance [@Haak2015; @Mathieson2015; @Rohland2022]. While established tools exist for this task for high-quality modern sequencing data [@samtools; @gatk], these are often not appropriate for ancient DNA, which has often too low sequencing-coverage and a higher error rate due to post-mortem DNA damage. PileupCaller is a command-line tool written in Haskell, which randomly samples genotypes from raw alignment data at predefined bi-allelic positions. Several modes can be selected, geared towards specific input data features and research questions.
+Next generation sequencing data is ubiquitous in medical and biological sciences. It has also become the primary tool in archaeogenetics, where ancient DNA is extracted from archaeological organic (often human skeletal) material, processed into DNA sequencing libraries and then sequenced [@Orlando2021]. As a testimony to the rapid and accelerating growth of the field, we today have close to ten thousand published ancient human genomes available in the public record [@Schmid2024; @Mallick2024], and many smaller datasets of other organisms. A key step in processing raw sequencing data is the estimation of genotypes at specific variable positions along the genome. Such positions are often pre-selected because they are informative about ancestry or of particular biological relevance [@Haak2015; @Mathieson2015; @Rohland2022]. While established tools exist for this task for high-quality modern sequencing data [@samtools; @gatk], these are often not appropriate for ancient DNA, which has often too low sequencing-coverage and a higher error rate due to post-mortem DNA damage. PileupCaller is a command-line tool written in Haskell, which randomly samples genotypes from raw alignment data at predefined bi-allelic positions. Several modes can be selected, geared towards specific input data features and research questions.
 
 # Statement of need
 
-Present-day DNA, for example from medical studies results in raw sequencing data with relatively low per-base error rates and sequencing-coverages of at least several multiples of 1 (for example [@1000_Genomes_Project_Consortium2015]) but in fact up to 20-30x coverage. Dedicated tools to process such data include samtools/bcftools [@samtools] and GATK [@gatk] among many other tools. Ancient DNA seuqencing data often comes with substantially lower coverage and substantially higher error rates. In terms of coverage, most ancient genomes have genome-wide coverage often below 1x and in fact very often even below 0.1x. Such low coverage means that any given genomic site is more likely not covered by a sequencing read than covered. At the same time, the low fraction of sites that is actually covered has higher error rates than modern DNA, due to ancient-DNA damage. These two factors violate the assumptions behind statistical genotype callers like `bcftools call` or `HaplotypeCaller` from GATK. 
[…]
