### 2026-04-01 #8634 @stschiff
Hi @Kevin-Mattheus-Moerman I have completed my ToDos:
- I have adapted the LICENSE text to be fully aligned with BSD-3
- I have added a `CONTRIBUTING.md` guide and a Code of Conduct.
- About this point: "Archive listed title and authors matches paper" -> This is correct and intentional. The contributions from the two additional authors are very minor.
- I fixed the two typos and indeed added an Acknowledgment section, thank you for the suggestion.
