### 2026-04-27 #8634 @mahfuz05062
> [@mahfuz05062](https://github.com/mahfuz05062) can you respond as well please?

I went over the paper and the repository. I think we are good, both functionality-wise and needed documentation for test and contributions. I marked those done on the reviewers checklist too!
