### 2026-05-02 #8634 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
