### 2026-05-02 #8634 @arfon
@stschiff – it looks like those typos have been fixed in your default branch. Is that where we should generate the paper from now (rather than the `writing_joss_paper` branch)
